+++ packages/atlas/src/views/studio/MyPaymentsView/PaymentsOverview/index.ts
@@ -1,0 +1,1 @@
+export * from './PaymentsOverView'
