--- packages/atlas/src/views/studio/MyPaymentsView/PaymentsOverviewTab/index.ts
@@ -1,1 +1,0 @@
-export * from './PaymentsOverviewTab'
