--- packages/atlas/src/views/studio/MyPaymentsView/PaymentsTransactionTab/index.ts
@@ -1,1 +1,0 @@
-export * from './PaymentTransactionsTab'
