+++ packages/atlas/src/views/studio/MyPaymentsView/PaymentsTransactions/index.ts
@@ -1,0 +1,1 @@
+export * from './PaymentTransactions'
